+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii EAuth Change Log
 ### In progress...
 * Added handling for denied callback in the TwitterOAuthService.
 * Fixed a redirect page for disabled javascript.
+* EAuthWidget been rewritten for use with CController->widget() instead of EAuth->renderWidget().
 
 ### Version 1.1.3 (14.10.2011)
 * MailruOAuthService::makeSignedRequest() now fully compatible with the basic method.